+++ CHANGELOG.md
@@ -4,10 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 ### Added
+ - API server
+ - API Cliente
  - Integra Transporter para [Fetch]
  - Integra Transporter para [Socket.io]
- - Incializa Appbase
-
+ - 💟 Incializa Appbase
 
 [Socket.io]: http://socket.io/docs/
 [Fetch]: https://developer.mozilla.org/en/docs/Web/API/Fetch_API